+++ objectbox/lib/src/native/query/builder.dart
@@ -13,10 +13,15 @@ class QueryBuilder<T> extends _QueryBuilder<T> {
 
   /// Finish building a [Query]. Call [Query.close()] after you're done with it
   /// to free resources.
+  ///
+  /// Can only be called once: the builder is closed afterwards, calling any
+  /// of its methods then throws a [StateError].
   Query<T> build() {
-    _applyCondition();
-
+    _checkNotClosed();
     try {
+      // Apply inside try: the native builder must be closed (freed) even if
+      // applying a condition fails.
+      _applyCondition();
       return Query<T>._(_store, _cBuilder, _entity);
     } finally {
       _close();
@@ -89,7 +94,7 @@ class QueryBuilder<T> extends _QueryBuilder<T> {
   ///     .build();
   /// ```
   QueryBuilder<T> order<D>(QueryProperty<T, D> p, {int flags = 0}) {
-    checkObx(C.qb_order(_cBuilder, p._model.id.id, flags));
+    checkObx(C.qb_order(_cBuilderChecked, p._model.id.id, flags));
     // Using Dart's cascade operator does not allow for nice chaining with
     // build(), so explicitly return this for a more fluent interface.
     // ignore: avoid_returning_this
@@ -104,6 +109,7 @@ class _QueryBuilder<T> {
   final Condition<T>? _queryCondition;
   final Pointer<OBX_query_builder> _cBuilder;
   final _innerQBs = <_QueryBuilder>[];
+  bool _closed = false;
 
   _QueryBuilder(
       this._store, this._entity, this._queryCondition, this._cBuilder) {
@@ -114,8 +120,10 @@ class _QueryBuilder<T> {
       : _store = srcQB._store,
         _entity = InternalStoreAccess.entityDef<T>(srcQB._store) {
     checkObxPtr(_cBuilder, 'failed to create QueryBuilder');
-    _applyCondition();
+    // Register before applying the condition so the native builder is closed
+    // with (and not leaked by) the source builder if applying throws.
     srcQB._innerQBs.add(this);
+    _applyCondition();
   }
 
   void _fillQueriedEntities(Set<Type> outEntities) {
@@ -126,10 +134,27 @@ class _QueryBuilder<T> {
   }
 
   void _close() {
+    if (_closed) return;
+    _closed = true;
     for (var iqb in _innerQBs) {
       iqb._close();
     }
     checkObx(C.qb_close(_cBuilder));
+  }
+
+  @pragma('vm:prefer-inline')
+  void _checkNotClosed() {
+    if (_closed) {
+      throw StateError('QueryBuilder is closed (build() was already called), '
+          'create a new one to build another query.');
+    }
+  }
+
+  /// [_cBuilder], but throws instead of using a freed native builder if this
+  /// was already closed.
+  Pointer<OBX_query_builder> get _cBuilderChecked {
+    _checkNotClosed();
+    return _cBuilder;
   }
 
   @pragma('vm:prefer-inline')
@@ -157,7 +182,7 @@ class _QueryBuilder<T> {
           QueryRelationToOne<T, TargetEntityT> rel,
           [Condition<TargetEntityT>? qc]) =>
       _QueryBuilder<TargetEntityT>._link(
-          this, qc, C.qb_link_property(_cBuilder, rel._model.id.id));
+          this, qc, C.qb_link_property(_cBuilderChecked, rel._model.id.id));
 
   /// Like [link], but where the to-one relation is defined in the other object.
   _QueryBuilder<SourceEntityT> backlink<SourceEntityT>(
@@ -167,7 +192,7 @@ class _QueryBuilder<T> {
           this,
           qc,
           C.qb_backlink_property(
-              _cBuilder,
+              _cBuilderChecked,
               InternalStoreAccess.entityDef<SourceEntityT>(_store).model.id.id,
               rel._model.id.id));
 
@@ -183,12 +208,12 @@ class _QueryBuilder<T> {
           QueryRelationToMany<T, TargetEntityT> rel,
           [Condition<TargetEntityT>? qc]) =>
       _QueryBuilder<TargetEntityT>._link(
-          this, qc, C.qb_link_standalone(_cBuilder, rel._model.id.id));
+          this, qc, C.qb_link_standalone(_cBuilderChecked, rel._model.id.id));
 
   /// Like [linkMany], but where the to-many relation is defined in the other object.
   _QueryBuilder<SourceEntityT> backlinkMany<SourceEntityT>(
           QueryRelationToMany<SourceEntityT, T> rel,
           [Condition<SourceEntityT>? qc]) =>
-      _QueryBuilder<SourceEntityT>._link(
-          this, qc, C.qb_backlink_standalone(_cBuilder, rel._model.id.id));
+      _QueryBuilder<SourceEntityT>._link(this, qc,
+          C.qb_backlink_standalone(_cBuilderChecked, rel._model.id.id));
 }